refactoring data operations (records) out into own file #49
